+++ src/main.ts
@@ -21,7 +21,7 @@ import { DotnetLoggerObserver } from './observers/DotnetLoggerObserver';
 import { OmnisharpDebugModeLoggerObserver } from './observers/OmnisharpDebugModeLoggerObserver';
 import { ActivationFailure, ActiveTextEditorChanged } from './omnisharp/loggingEvents';
 import { EventStream } from './EventStream';
-import { OmnisharpServerStatusObserver, ExecuteCommand, ShowWarningMessage, MessageItemWithCommand } from './observers/OmnisharpServerStatusObserver';
+import { WarningMessageObserver, ExecuteCommand, ShowWarningMessage, MessageItemWithCommand } from './observers/WarningMessageObserver';
 import { InformationMessageObserver, ShowInformationMessage, GetConfiguration, WorkspaceAsRelativePath } from './observers/InformationMessageObserver';
 import { GetActiveTextEditor, OmnisharpStatusBarObserver, Match } from './observers/OmnisharpStatusBarObserver';
 import { TextEditorAdapter } from './textEditorAdapter';
@@ -59,7 +59,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<{ init
 
     let showWarningMessage: ShowWarningMessage<MessageItemWithCommand> = <T extends vscode.MessageItem>(message: string, ...items: T[]) => vscode.window.showWarningMessage(message, ...items);
     let executeCommand: ExecuteCommand<string> = <T>(command: string, ...rest: any[]) => vscode.commands.executeCommand(command, ...rest);
-    let omnisharpServerStatusObserver = new OmnisharpServerStatusObserver(showWarningMessage, executeCommand);
+    let omnisharpServerStatusObserver = new WarningMessageObserver(showWarningMessage, executeCommand);
     eventStream.subscribe(omnisharpServerStatusObserver.post);
 
     let getConfiguration: GetConfiguration = (name: string) => vscode.workspace.getConfiguration(name);
